+++ plugins/simulator/waveform_viewer/include/waveform_viewer/wave_graphics_canvas.h
@@ -25,11 +25,11 @@ namespace hal {
         WaveTimescale* mTimescale;
         WaveCursor* mCursor;
         bool mMoveCursor;
-        float mCursorTime;
+        double mCursorTime;
         int mCursorXpos;
         QWidget* mDragZoom;
     Q_SIGNALS:
-        void cursorMoved(float tCursor, int xpos);
+        void cursorMoved(double tCursor, int xpos);
     public Q_SLOTS:
         void updateRequest();
         void handleWaveUpdated(int iwave, int groupId);
@@ -47,6 +47,6 @@ namespace hal {
         const WaveTransform* transform() const { return &mTransform; }
         const WaveItemHash* waveItemHash() const { return mWaveItemHash; }
         void toggleZoom();
-        void setCursorPosition(float tCursor, int xpos);
+        void setCursorPosition(double tCursor, int xpos);
     };
 }
